+++ src/main/java/io/github/mzmine/modules/dataprocessing/id_ccscalibration/external/WatersImsCalibrationReader.java
@@ -1,5 +1,25 @@
-package io.github.mzmine.modules.dataprocessing.id_ccscalibration;
+/*
+ *  Copyright 2006-2020 The MZmine Development Team
+ *
+ *  This file is part of MZmine.
+ *
+ *  MZmine is free software; you can redistribute it and/or modify it under the terms of the GNU
+ *  General Public License as published by the Free Software Foundation; either version 2 of the
+ *  License, or (at your option) any later version.
+ *
+ *  MZmine is distributed in the hope that it will be useful, but WITHOUT ANY WARRANTY; without even
+ *  the implied warranty of MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the GNU General
+ *  Public License for more details.
+ *
+ *  You should have received a copy of the GNU General Public License along with MZmine; if not,
+ *  write to the Free Software Foundation, Inc., 51 Franklin St, Fifth Floor, Boston, MA 02110-1301
+ *  USA
+ */
 
+package io.github.mzmine.modules.dataprocessing.id_ccscalibration.external;
+
+import io.github.mzmine.modules.dataprocessing.id_ccscalibration.CCSCalibration;
+import io.github.mzmine.modules.dataprocessing.id_ccscalibration.TwCCSCalibration;
 import java.io.BufferedReader;
 import java.io.File;
 import java.io.FileReader;
@@ -12,6 +32,9 @@ import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
 
 public class WatersImsCalibrationReader {
+
+  private WatersImsCalibrationReader() {
+  }
 
   private static final Logger logger = Logger.getLogger(WatersImsCalibrationReader.class.getName());
 
@@ -74,7 +97,7 @@ public class WatersImsCalibrationReader {
     final double exp = Double.parseDouble(strExponent);
     final double t0 = Double.parseDouble(strT0);
 
-    final File externInf = new File(file.getParentFile(), "_extern.inf");
+    final File externInf = new File(file.getParentFile(), WATERS_EXTERN_INF);
 
     if (!file.exists() || !file.canRead()) {
       throw new IllegalArgumentException("Cannot find or read extern.inf file.");
